# backend/api/swagger.yaml
openapi: 3.0.0
info:
  title: Auto Industry Marketplace API
  version: 1.0.0
  description: API documentation for the Auto Industry Marketplace platform
  contact:
    name: API Support
    email: support@automarketplace.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.automarketplace.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [car_owner, repair_shop, vendor]
        businessName:
          type: string
        businessLicense:
          type: string
        phone:
          type: string
        address:
          type: object
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    SparePart:
      type: object
      properties:
        id:
          type: integer
        vendor_id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        description:
          type: string
        sku:
          type: string
        compatible_cars:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - role
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
                role:
                  type: string
                  enum: [car_owner, repair_shop, vendor]
                phone:
                  type: string
                address:
                  type: object
                businessName:
                  type: string
                  description: Required for repair_shop and vendor roles
                businessLicense:
                  type: string
                  description: Required for repair_shop and vendor roles
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        409:
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Account is inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Logged out successfully
...
  /api/parts:
    get:
      tags:
        - Spare Parts
      summary: Get all spare parts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: List of spare parts
          content:
            application/json:
              schema:
                type: object
                properties:
                  parts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SparePart'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

    post:
      tags:
        - Spare Parts
      summary: Add new spare part
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - category
                - price
                - stock
              properties:
                name:
                  type: string
                category:
                  type: string
                price:
                  type: number
                  minimum: 0
                stock:
                  type: integer
                  minimum: 0
                description:
                  type: string
                sku:
                  type: string
                compatibleCars:
                  type: array
                  items:
                    type: string
      responses:
        201:
          description: Part added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  part:
                    $ref: '#/components/schemas/SparePart'
  /api/parts/{id}:
    get:
      tags:
        - Spare Parts
      summary: Get single spare part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Spare part details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SparePart'
        404:
          description: Part not found

    put:
      tags:
        - Spare Parts
      summary: Update spare part
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - category
                - price
                - stock
              properties:
                name:
                  type: string
                category:
                  type: string
                price:
                  type: number
                stock:
                  type: integer
                description:
                  type: string
                sku:
                  type: string
                compatibleCars:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: Part updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  part:
                    $ref: '#/components/schemas/SparePart'
  /api/parts/{id}/stock:
    patch:
      tags:
        - Spare Parts
      summary: Update part stock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stock
              properties:
                stock:
                  type: integer
                operation:
                  type: string
                  enum: [set, increment, decrement]
                  default: set
      responses:
        200:
          description: Stock updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  part:
                    $ref: '#/components/schemas/SparePart'
security:
  - bearerAuth: []
...